backend/requirements.txt
django==4.2.1
django-environ==0.10.0
django-cors-headers==3.14.0

djangorestframework==3.14.0

psycopg2==2.9.6

Pillow==9.5.0

stripe==5.4.0
